Body parsing

+++ lib/jb/src/features/int/my_strlen.c
@@ -7,7 +7,7 @@
 
 #include "jb.h"
 
-size_t my_strlen(char const *str)
+int my_strlen(char const *str)
 {
     size_t i = 0;
